+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
   * raised minimum required PHP version to 7.0.0
   * introduced scalar type hints and strict type checking
+  * renamed `stubbles\input\AbstractRequest` to `stubbles\input\ParamRequest`
   * removed methods deprecated in 6.0.0
     * `stubbles\input\ValueReader::asSecureString()`, use `stubbles\input\ValueReader::asSecret()` instead
     * `stubbles\input\ValueReader::ifSatisfiesRegex()`, use `stubbles\input\ValueReader::ifMatches()` instead